Rearranging some of the abstract classes
