+++ project/hca/indexer.py
@@ -1,3 +1,5 @@
 from utils.indexer import Indexer
 from project.hca.transformers import FileTransformer
 
+class HCAIndexer(Indexer):
+    pass
